requirements.txt
django==5.0.6
djangorestframework==3.15.2
psycopg[binary]==3.2.13
django-environ==0.11.2
gunicorn==22.0.0
whitenoise==6.7.0
firebase-admin==6.5.0
pytest==8.3.2
pytest-django==4.8.0
google-generativeai==0.8.4
